refactor: 도메인 종속되지 않은 공통 컴포넌트 폴더 분리
