+++ src/components/@common/avatar/Avatar.stories.ts
